requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.9.1
asgiref==3.4.1
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==21.12b0
blinker==1.4
ccxt==2.0.20
certifi==2023.5.7
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.3
cloud-init==20.1
colorama==0.4.3
colorlog==6.7.0
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==36.0.2
dbus-python==1.2.16
decorator==5.1.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.12
django-apscheduler==0.6.2
django-simpleui==2022.1
entrypoints==0.3
exceptiongroup==1.1.1
executing==0.8.2
filelock==3.4.2
frozenlist==1.3.1
gunicorn==20.1.0
h11==0.14.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipython==8.0.0
jedi==0.18.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.2
MarkupSafe==1.1.0
matplotlib-inline==0.1.3
more-itertools==4.2.0
multidict==6.0.2
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.24.3
oauthlib==3.1.0
outcome==1.2.0
pandas==2.0.1
parso==0.8.3
pathspec==0.9.0
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycares==4.2.2
pycparser==2.21
pycryptodome==3.17
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.9.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
ssh-import-id==5.10
stack-data==0.1.3
stevedore==3.5.0
systemd-python==234
tomli==1.2.3
traitlets==5.1.1
trio==0.22.0
trio-websocket==0.10.2
Twisted==18.9.0
typing-extensions==4.1.1
tzdata==2022.5
tzlocal==4.2
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==2.0.2
virtualenv==20.13.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
wsproto==1.2.0
yarl==1.8.1
zipp==1.0.0
zope.interface==4.7.1
